basket/requirements.txt
flask~=2.2.3
python-dotenv
kafka-python